server/requirements.txt
Django==1.9.1
Pygments==2.1
blessings==1.6
bpython==0.15
curtsies==0.2.6
distribute==0.6.24
djangorestframework==3.3.2
greenlet==0.4.9
requests==2.9.1
six==1.10.0
wcwidth==0.1.6
wsgiref==0.1.2
